+++ src/Models/order/Transformers/OrderTransformer.ts
@@ -2,11 +2,13 @@ import Order from '../Order';
 import OrderInterface from '../../../Interfaces/order/Order';
 import OrderLineInterface from '../../../Interfaces/order/OrderLine';
 import UserInterface from '../../../Interfaces/user/User';
+import AssigneeInterface from '../../../Interfaces/order/Assignee';
+import { Basket as BasketInterface } from '../../../Interfaces/basket/Basket';
 import UserTransformer from '../../user/Transformers/UserTransformer';
 import OrderLineTransformer from '../../order/Transformers/OrderLineTransformer';
 import BaseTransformer from '../../../BaseTransformer';
-import AssigneeInterface from '../../../Interfaces/order/Assignee';
 import AssigneeTransformer from './AssigneeTransformer';
+import BasketTransformer from '../../basket/Transformers/BasketTransformer';
 
 export default class OrderTransformer extends BaseTransformer {
     /**
@@ -62,10 +64,15 @@ export default class OrderTransformer extends BaseTransformer {
             expectedDeliveryDate: order.expectedDeliveryDate,
             hasReusablePackagingOptIn: order.hasReusablePackagingOptIn,
             weight: order.weight,
+            shippingServiceId: order.shippingServiceId,
             // belongs to
             user: this.includeUser(order),
+            basket: this.includeBasket(order),
+            assignee: this.includeAssignee(order),
             // has many
             lines: this.includeOrderLines(order),
+            // accessors
+            isGift: order.isGift,
         });
     }
 
@@ -73,11 +80,15 @@ export default class OrderTransformer extends BaseTransformer {
         return this.item(order, 'user', new UserTransformer());
     }
 
-    includeOrderLines(order: OrderInterface): OrderLineInterface[] {
+    includeOrderLines(order: OrderInterface): Array<OrderLineInterface> {
         return this.collection(order, 'lines', new OrderLineTransformer());
     }
 
     includeAssignee(order: OrderInterface): AssigneeInterface | null {
         return this.item(order, 'assignee', new AssigneeTransformer());
+    }
+
+    includeBasket(order: OrderInterface): BasketInterface | null {
+        return this.item(order, 'basket', new BasketTransformer());
     }
 }